src/requirements.txt
discord.py[voice]
cogs
requests
asyncio
